pyproject.toml
[tool.poetry]
name = "img"
version = "0.1.0"
description = '''

Простой скрипт, конвертирующий JPG картинки в PNG. Данный скрипт принмает два аргумента из командной строки:
В первом аргументе передаём путь до папки с изображениями, которые мы хотим конвертировать.
Во втором аргументе передаём название папки, в которую будут сохранены изображения.
Важно! Чтобы скрипт работал корректно, нужно, чтобы папка с изображениями находилась рядом со скриптом.

'''
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
